gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = ""

readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "grpcio (>=1.75.1,<2.0.0)",
    "grpcio-tools (>=1.75.1,<2.0.0)",
    "pydantic (>=2.12.2,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
